feat: Section 18 — amount 단위 BigDecimal→BigInteger 전환 및 테스트 픽스
- CreateWithdrawalRequest.amount를 ETH 단위 BigDecimal에서 최소 단위 BigInteger(wei)로 변경
- AmountLimitPolicyRule maxAmount 타입 BigDecimal→BigInteger 전환
- policy.max-amount 테스트 기본값을 wei 기준 대형 값으로 조정 (9999999999999999999)
- 통합 테스트 amount 필드를 wei 단위 정수로 일괄 변환 (0.1 ETH→100000000000000000 등)
- data.sql 추가: H2 테스트 시드 데이터 (supported_chains/assets, MERGE INTO 구문)
- spring.jpa.defer-datasource-initialization=true로 data.sql 실행 타이밍 보장

+++ custody/src/test/java/lab/custody/orchestration/policy/ToAddressWhitelistPolicyRuleTest.java
@@ -10,7 +10,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -122,6 +122,6 @@ class ToAddressWhitelistPolicyRuleTest {
                 ? "0x0000000000000000000000000000000000000000"
                 : toAddress;
         return new CreateWithdrawalRequest("evm", "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
-                addr, "ETH", new BigDecimal("1"));
+                addr, "ETH", new BigInteger("1"));
     }
 }
